requirements.txt
openai>=1.0.0
anthropic>=0.39.0
google-genai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
requests>=2.31.0
openpyxl>=3.1.0
google-cloud-storage>=2.0.0
